pyproject.toml
[tool.poetry]
name = "zelenkaparser"
version = "0.1.0"
description = ""
authors = ["hachimaru <74972395+xxspell@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.1"
structlog = "^24.1.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
httpx-html = "^0.11.0.dev0"
selenium = "^4.16.0"
undetected-chromedriver = "^3.5.4"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
